+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Bumped up PromptingTools to v0.20 (brings new RAG capabilities, pretty-printing, etc.)
 - Changed default model to be GPT-4 Turbo to improve answer quality
+- Documentation moved to Vitepress
 
 ### Fixed
 - Fixed wrong initiation of `CONV_HISTORY` and other globals that led to UndefVarError. Moved several globals to `const Ref{}` pattern to ensure type stability, but it means that from now it always needs to be dereferenced with `[]` (eg, `MAIN_INDEX[]` instead of `MAIN_INDEX`).